Rename subjects package to subject
