+++ examples/timeflies/timeflies_gtk.py
@@ -1,6 +1,6 @@
 import rx
 from rx import operators as ops
-from rx.subjects import Subject
+from rx.subject import Subject
 from rx.scheduler.mainloop import GtkScheduler
 
 import gi
