+++ examples/timeflies/timeflies_qt.py
@@ -2,7 +2,7 @@ import sys
 
 import rx
 from rx import operators as ops
-from rx.subjects import Subject
+from rx.subject import Subject
 from rx.scheduler.mainloop import QtScheduler
 
 try:
