+++ rx/core/operators/windowwithcount.py
@@ -5,7 +5,7 @@ from rx.core import Observable
 from rx.internal.utils import add_ref
 from rx.disposable import SingleAssignmentDisposable, RefCountDisposable
 from rx.internal.exceptions import ArgumentOutOfRangeException
-from rx.subjects import Subject
+from rx.subject import Subject
 
 log = logging.getLogger("Rx")
