+++ rx/core/operators/windowwithtimeorcount.py
@@ -5,7 +5,7 @@ from rx.core import Observable, typing
 from rx.scheduler import timeout_scheduler
 from rx.internal.utils import add_ref
 from rx.disposable import SingleAssignmentDisposable, CompositeDisposable, RefCountDisposable, SerialDisposable
-from rx.subjects import Subject
+from rx.subject import Subject
 
 
 def _window_with_time_or_count(timespan: typing.RelativeTime, count: int, scheduler: Optional[typing.Scheduler] = None
