+++ tests/test_subject/test_asyncsubject.py
@@ -1,7 +1,7 @@
 import pytest
 
 from rx.testing import TestScheduler, ReactiveTest
-from rx.subjects import BehaviorSubject
+from rx.subject import AsyncSubject
 from rx.internal.exceptions import DisposedException
 
 on_next = ReactiveTest.on_next
@@ -23,8 +23,13 @@ def _raise(ex):
 
 
 def test_infinite():
-    scheduler = TestScheduler()
+    subject = [None]
+    subscription = [None]
+    subscription1 = [None]
+    subscription2 = [None]
+    subscription3 = [None]
 
+    scheduler = TestScheduler()
     xs = scheduler.create_hot_observable(
         on_next(70, 1),
         on_next(110, 2),
@@ -39,19 +44,12 @@ def test_infinite():
         on_next(940, 11),
         on_next(1020, 12)
     )
-
-    subject = [None]
-    subscription = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
-
     results1 = scheduler.create_observer()
     results2 = scheduler.create_observer()
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        subject[0] = AsyncSubject()
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -91,27 +89,19 @@ def test_infinite():
     scheduler.schedule_absolute(950, action10)
 
     scheduler.start()
-
-    assert results1.messages == [
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)]
-
-    assert results2.messages == [
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_next(630, 8)]
-
-    assert results3.messages == [
-        on_next(900, 10),
-        on_next(940, 11)]
+    assert results1.messages == []
+    assert results2.messages == []
+    assert results3.messages == []
 
 
 def test_finite():
-    scheduler = TestScheduler()
+    subject = [None]
+    subscription = [None]
+    subscription1 = [None]
+    subscription2 = [None]
+    subscription3 = [None]
 
+    scheduler = TestScheduler()
     xs = scheduler.create_hot_observable(
         on_next(70, 1),
         on_next(110, 2),
@@ -123,21 +113,14 @@ def test_finite():
         on_completed(630),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException())
+        on_error(660, 'ex')
     )
-
-    subject = [None]
-    subscription = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
-
     results1 = scheduler.create_observer()
     results2 = scheduler.create_observer()
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        subject[0] = AsyncSubject()
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -177,28 +160,19 @@ def test_finite():
     scheduler.schedule_absolute(950, action10)
 
     scheduler.start()
-
-    assert results1.messages == [
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)]
-
-    assert results2.messages == [
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_completed(630)]
-
-    assert results3.messages == [
-        on_completed(900)]
-
+    assert results1.messages == []
+    assert results2.messages == [on_next(630, 7), on_completed(630)]
+    assert results3.messages == [on_next(900, 7), on_completed(900)]
 
 def test_error():
+    subject = [None]
+    subscription = [None]
+    subscription1 = [None]
+    subscription2 = [None]
+    subscription3 = [None]
+
+    ex = 'ex'
     scheduler = TestScheduler()
-
-    ex = RxException()
-
     xs = scheduler.create_hot_observable(
         on_next(70, 1),
         on_next(110, 2),
@@ -210,98 +184,79 @@ def test_error():
         on_error(630, ex),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException())
+        on_error(660, 'ex2')
     )
+    results1 = scheduler.create_observer()
+    results2 = scheduler.create_observer()
+    results3 = scheduler.create_observer()
 
+    def action(scheduler, state=None):
+        subject[0] = AsyncSubject()
+    scheduler.schedule_absolute(100, action)
+
+    def action1(scheduler, state=None):
+        subscription[0] = xs.subscribe(subject[0])
+    scheduler.schedule_absolute(200, action1)
+
+    def action2(scheduler, state=None):
+        subscription[0].dispose()
+    scheduler.schedule_absolute(1000, action2)
+
+    def action3(scheduler, state=None):
+        subscription1[0] = subject[0].subscribe(results1)
+    scheduler.schedule_absolute(300, action3)
+
+    def action4(scheduler, state=None):
+        subscription2[0] = subject[0].subscribe(results2)
+    scheduler.schedule_absolute(400, action4)
+
+    def action5(scheduler, state=None):
+        subscription3[0] = subject[0].subscribe(results3)
+    scheduler.schedule_absolute(900, action5)
+
+    def action6(scheduler, state=None):
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(600, action6)
+
+    def action7(scheduler, state=None):
+        subscription2[0].dispose()
+    scheduler.schedule_absolute(700, action7)
+
+    def action8(scheduler, state=None):
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(800, action8)
+
+    def action9(scheduler, state=None):
+        subscription3[0].dispose()
+    scheduler.schedule_absolute(950, action9)
+
+    scheduler.start()
+    assert results1.messages == []
+    assert results2.messages == [on_error(630, ex)]
+    assert results3.messages == [on_error(900, ex)]
+
+
+def test_canceled():
     subject = [None]
     subscription = [None]
     subscription1 = [None]
     subscription2 = [None]
     subscription3 = [None]
 
-    results1 = scheduler.create_observer()
-    results2 = scheduler.create_observer()
-    results3 = scheduler.create_observer()
-
-    def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
-    scheduler.schedule_absolute(100, action1)
-
-    def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action2)
-
-    def action3(scheduler, state=None):
-        subscription[0].dispose()
-    scheduler.schedule_absolute(1000, action3)
-
-    def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action4)
-
-    def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action5)
-
-    def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
-    scheduler.schedule_absolute(900, action6)
-
-    def action7(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(600, action7)
-
-    def action8(scheduler, state=None):
-        subscription2[0].dispose()
-    scheduler.schedule_absolute(700, action8)
-
-    def action9(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(800, action9)
-
-    def action10(scheduler, state=None):
-        subscription3[0].dispose()
-    scheduler.schedule_absolute(950, action10)
-
-    scheduler.start()
-
-    assert results1.messages == [
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)]
-
-    assert results2.messages == [
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_error(630, ex)]
-
-    assert results3.messages == [
-        on_error(900, ex)]
-
-def test_canceled():
     scheduler = TestScheduler()
-
     xs = scheduler.create_hot_observable(
         on_completed(630),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException())
-    )
-
-    subject = [None]
-    subscription = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
+        on_error(660, 'ex')
+        )
 
     results1 = scheduler.create_observer()
     results2 = scheduler.create_observer()
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(100)
+        subject[0] = AsyncSubject()
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -341,33 +296,23 @@ def test_canceled():
     scheduler.schedule_absolute(950, action10)
 
     scheduler.start()
-
-    assert results1.messages == [
-        on_next(300, 100)]
-
-    assert results2.messages == [
-        on_next(400, 100),
-        on_completed(630)]
-
-    assert results3.messages == [
-        on_completed(900)]
+    assert results1.messages == []
+    assert results2.messages == [on_completed(630)]
+    assert results3.messages == [on_completed(900)]
 
 def test_subject_disposed():
+    subject = [None]
+    subscription1 = [None]
+    subscription2 = [None]
+    subscription3 = [None]
     scheduler = TestScheduler()
 
-    subject = [None]
-
     results1 = scheduler.create_observer()
-    subscription1 = [None]
-
     results2 = scheduler.create_observer()
-    subscription2 = [None]
-
     results3 = scheduler.create_observer()
-    subscription3 = [None]
 
     def action1(scheduler, state=None):
-        subject[0] = BehaviorSubject(0)
+        subject[0] = AsyncSubject()
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -430,7 +375,7 @@ def test_subject_disposed():
 
     def action16(scheduler, state=None):
         with pytest.raises(DisposedException):
-            subject[0].on_error(RxException())
+            subject[0].on_error('ex')
     scheduler.schedule_absolute(850, action16)
 
     def action17(scheduler, state=None):
@@ -439,20 +384,9 @@ def test_subject_disposed():
     scheduler.schedule_absolute(950, action17)
 
     scheduler.start()
+    assert results1.messages == []
+    assert results2.messages == []
+    assert results3.messages == []
 
-    assert results1.messages == [
-        on_next(200, 1),
-        on_next(250, 2),
-        on_next(350, 3),
-        on_next(450, 4)]
-
-    assert results2.messages == [
-        on_next(300, 2),
-        on_next(350, 3),
-        on_next(450, 4),
-        on_next(550, 5)]
-
-    assert results3.messages == [
-        on_next(400, 3),
-        on_next(450, 4),
-        on_next(550, 5)]
+if __name__ == '__main__':
+    unittest.main()
