+++ tests/test_subject/test_behaviorsubject.py
@@ -1,8 +1,7 @@
-import sys
 import pytest
 
 from rx.testing import TestScheduler, ReactiveTest
-from rx.subjects import ReplaySubject
+from rx.subject import BehaviorSubject
 from rx.internal.exceptions import DisposedException
 
 on_next = ReactiveTest.on_next
@@ -52,97 +51,7 @@ def test_infinite():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
-    scheduler.schedule_absolute(100, action1)
-
-    def action2(scheduler, state=None):
-        subscription[0] = xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action2)
-
-    def action3(scheduler, state=None):
-        subscription[0].dispose()
-    scheduler.schedule_absolute(1000, action3)
-
-    def action4(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action4)
-
-    def action5(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action5)
-
-    def action6(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
-    scheduler.schedule_absolute(900, action6)
-
-    def action7(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(600, action7)
-
-    def action8(scheduler, state=None):
-        subscription2[0].dispose()
-    scheduler.schedule_absolute(700, action8)
-
-    def action9(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(800, action9)
-
-    def action10(scheduler, state=None):
-        subscription3[0].dispose()
-    scheduler.schedule_absolute(950, action10)
-
-    scheduler.start()
-
-    assert results1.messages == [
-        on_next(300, 3),
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7)]
-
-    assert results2.messages == [
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_next(630, 8)]
-
-    assert results3.messages == [
-        on_next(900, 10),
-        on_next(940, 11)]
-
-
-def test_infinite2():
-    scheduler = TestScheduler()
-
-    xs = scheduler.create_hot_observable(
-        on_next(70, 1),
-        on_next(110, 2),
-        on_next(220, 3),
-        on_next(270, 4),
-        on_next(280, -1),
-        on_next(290, -2),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_next(630, 8),
-        on_next(710, 9),
-        on_next(870, 10),
-        on_next(940, 11),
-        on_next(1020, 12)
-    )
-
-    subject = [None]
-    subscription = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
-
-    results1 = scheduler.create_observer()
-    results2 = scheduler.create_observer()
-    results3 = scheduler.create_observer()
-
-    def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        subject[0] = BehaviorSubject(100)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -185,8 +94,6 @@ def test_infinite2():
 
     assert results1.messages == [
         on_next(300, 4),
-        on_next(300, -1),
-        on_next(300, -2),
         on_next(340, 5),
         on_next(410, 6),
         on_next(520, 7)]
@@ -200,6 +107,7 @@ def test_infinite2():
     assert results3.messages == [
         on_next(900, 10),
         on_next(940, 11)]
+
 
 def test_finite():
     scheduler = TestScheduler()
@@ -215,7 +123,7 @@ def test_finite():
         on_completed(630),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, 'ex')
+        on_error(660, RxException())
     )
 
     subject = [None]
@@ -229,49 +137,48 @@ def test_finite():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        subject[0] = BehaviorSubject(100)
     scheduler.schedule_absolute(100, action1)
 
-    def action3(scheduler, state=None):
+    def action2(scheduler, state=None):
         subscription[0] = xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action3)
+    scheduler.schedule_absolute(200, action2)
+
+    def action3(scheduler, state=None):
+        subscription[0].dispose()
+    scheduler.schedule_absolute(1000, action3)
 
     def action4(scheduler, state=None):
-        subscription[0].dispose()
-    scheduler.schedule_absolute(1000, action4)
+        subscription1[0] = subject[0].subscribe(results1)
+    scheduler.schedule_absolute(300, action4)
 
     def action5(scheduler, state=None):
-        subscription1[0] = subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action5)
+        subscription2[0] = subject[0].subscribe(results2)
+    scheduler.schedule_absolute(400, action5)
 
     def action6(scheduler, state=None):
-        subscription2[0] = subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action6)
+        subscription3[0] = subject[0].subscribe(results3)
+    scheduler.schedule_absolute(900, action6)
 
     def action7(scheduler, state=None):
-        subscription3[0] = subject[0].subscribe(results3)
-    scheduler.schedule_absolute(900, action7)
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(600, action7)
 
     def action8(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(600, action8)
+        subscription2[0].dispose()
+    scheduler.schedule_absolute(700, action8)
 
     def action9(scheduler, state=None):
-        subscription2[0].dispose()
-    scheduler.schedule_absolute(700, action9)
+        subscription1[0].dispose()
+    scheduler.schedule_absolute(800, action9)
 
     def action10(scheduler, state=None):
-        subscription1[0].dispose()
-    scheduler.schedule_absolute(800, action10)
-
-    def action11(scheduler, state=None):
         subscription3[0].dispose()
-    scheduler.schedule_absolute(950, action11)
+    scheduler.schedule_absolute(950, action10)
 
     scheduler.start()
 
     assert results1.messages == [
-        on_next(300, 3),
         on_next(300, 4),
         on_next(340, 5),
         on_next(410, 6),
@@ -290,7 +197,7 @@ def test_finite():
 def test_error():
     scheduler = TestScheduler()
 
-    ex = RxException('ex')
+    ex = RxException()
 
     xs = scheduler.create_hot_observable(
         on_next(70, 1),
@@ -303,7 +210,7 @@ def test_error():
         on_error(630, ex),
         on_next(640, 9),
         on_completed(650),
-        on_error(660, RxException('ex'))
+        on_error(660, RxException())
     )
 
     subject = [None]
@@ -317,7 +224,7 @@ def test_error():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        subject[0] = BehaviorSubject(100)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -359,7 +266,6 @@ def test_error():
     scheduler.start()
 
     assert results1.messages == [
-        on_next(300, 3),
         on_next(300, 4),
         on_next(340, 5),
         on_next(410, 6),
@@ -373,7 +279,6 @@ def test_error():
 
     assert results3.messages == [
         on_error(900, ex)]
-
 
 def test_canceled():
     scheduler = TestScheduler()
@@ -396,7 +301,7 @@ def test_canceled():
     results3 = scheduler.create_observer()
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(3, 100, scheduler)
+        subject[0] = BehaviorSubject(100)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -437,9 +342,11 @@ def test_canceled():
 
     scheduler.start()
 
-    assert results1.messages == []
+    assert results1.messages == [
+        on_next(300, 100)]
 
     assert results2.messages == [
+        on_next(400, 100),
         on_completed(630)]
 
     assert results3.messages == [
@@ -449,16 +356,18 @@ def test_subject_disposed():
     scheduler = TestScheduler()
 
     subject = [None]
-    subscription1 = [None]
-    subscription2 = [None]
-    subscription3 = [None]
 
     results1 = scheduler.create_observer()
+    subscription1 = [None]
+
     results2 = scheduler.create_observer()
+    subscription2 = [None]
+
     results3 = scheduler.create_observer()
+    subscription3 = [None]
 
     def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(scheduler=scheduler)
+        subject[0] = BehaviorSubject(0)
     scheduler.schedule_absolute(100, action1)
 
     def action2(scheduler, state=None):
@@ -521,14 +430,12 @@ def test_subject_disposed():
 
     def action16(scheduler, state=None):
         with pytest.raises(DisposedException):
-            subject[0].on_error(Exception())
-
+            subject[0].on_error(RxException())
     scheduler.schedule_absolute(850, action16)
 
     def action17(scheduler, state=None):
         with pytest.raises(DisposedException):
             subject[0].subscribe(None)
-
     scheduler.schedule_absolute(950, action17)
 
     scheduler.start()
@@ -540,85 +447,12 @@ def test_subject_disposed():
         on_next(450, 4)]
 
     assert results2.messages == [
-        on_next(300, 1),
         on_next(300, 2),
         on_next(350, 3),
         on_next(450, 4),
         on_next(550, 5)]
 
     assert results3.messages == [
-        on_next(400, 1),
-        on_next(400, 2),
         on_next(400, 3),
         on_next(450, 4),
         on_next(550, 5)]
-
-
-def test_replay_subject_dies_out():
-    scheduler = TestScheduler()
-
-    xs = scheduler.create_hot_observable(
-        on_next(70, 1),
-        on_next(110, 2),
-        on_next(220, 3),
-        on_next(270, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_completed(580)
-    )
-
-    subject = [None]
-
-    results1 = scheduler.create_observer()
-    results2 = scheduler.create_observer()
-    results3 = scheduler.create_observer()
-    results4 = scheduler.create_observer()
-
-    def action1(scheduler, state=None):
-        subject[0] = ReplaySubject(sys.maxsize, 100, scheduler)
-    scheduler.schedule_absolute(100, action1)
-
-    def action2(scheduler, state=None):
-        xs.subscribe(subject[0])
-    scheduler.schedule_absolute(200, action2)
-
-    def action3(scheduler, state=None):
-        subject[0].subscribe(results1)
-    scheduler.schedule_absolute(300, action3)
-
-    def action4(scheduler, state=None):
-        subject[0].subscribe(results2)
-    scheduler.schedule_absolute(400, action4)
-
-    def action5(scheduler, state=None):
-        subject[0].subscribe(results3)
-    scheduler.schedule_absolute(600, action5)
-
-    def action6(scheduler, state=None):
-        subject[0].subscribe(results4)
-    scheduler.schedule_absolute(900, action6)
-
-    scheduler.start()
-
-    assert results1.messages == [
-        on_next(300, 3),
-        on_next(300, 4),
-        on_next(340, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_completed(580)]
-
-    assert results2.messages == [
-        on_next(400, 5),
-        on_next(410, 6),
-        on_next(520, 7),
-        on_completed(580)]
-
-    assert results3.messages == [
-        on_next(600, 7),
-        on_completed(600)]
-
-    assert results4.messages == [
-        on_completed(900)]
-
